sitne izmene

+++ src/main/java/com/iktpreobuka/repositories/PredmetRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.iktpreobuka.entites.Predmet;
 
-public interface PredmetRepository extends CrudRepository<Predmet, Integer> {
+public interface PredmetRepository extends CrudRepository<Predmet, Long> {
 
 	Optional<Predmet> findByNameIgnoreCase(String name);
 }
